+++ oneview/data_source_interconnects.go
@@ -161,7 +161,7 @@ func dataSourceInterconnects() *schema.Resource {
 							Elem:     &schema.Schema{Type: schema.TypeString},
 							Set:      schema.HashString,
 						},
-						"trap_destination": {
+						"trap_destinations": {
 							Type:     schema.TypeList,
 							Computed: true,
 							Elem: &schema.Resource{
@@ -278,6 +278,7 @@ func dataSourceInterconnectsRead(d *schema.ResourceData, meta interface{}) error
 	interconnectLocation = append(interconnectLocation, map[string]interface{}{
 		"location_entries": locationEntries,
 	})
+
 	d.Set("interconnect_location", interconnectLocation)
 
 	ipAddressList := make([]map[string]interface{}, 0,
@@ -332,11 +333,15 @@ func dataSourceInterconnectsRead(d *schema.ResourceData, meta interface{}) error
 
 	snmpConfiguration := make([]map[string]interface{}, 0, 1)
 	snmpConfiguration = append(snmpConfiguration, map[string]interface{}{
-		"enabled":          *interconnect.SnmpConfiguration.Enabled,
-		"v3_enabled":       *interconnect.SnmpConfiguration.V3Enabled,
-		"snmp_access":      schema.NewSet(schema.HashString, snmpAccess),
-		"type":             interconnect.SnmpConfiguration.Type,
-		"trap_destination": trapDestinations,
+		"category":          interconnect.SnmpConfiguration.Category,
+		"description":       interconnect.SnmpConfiguration.Description,
+		"name":              interconnect.SnmpConfiguration.Name,
+		"uri":               interconnect.SnmpConfiguration.URI,
+		"enabled":           interconnect.SnmpConfiguration.Enabled,
+		"v3_enabled":        interconnect.SnmpConfiguration.V3Enabled,
+		"snmp_access":       schema.NewSet(schema.HashString, snmpAccess),
+		"type":              interconnect.SnmpConfiguration.Type,
+		"trap_destinations": trapDestinations,
 	})
 	d.Set("snmp_configuration", snmpConfiguration)