+++ CHANGELOG.md
@@ -12,7 +12,9 @@ This project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html
 - Uplink Set
 
 ### Major changes:
-- Refactored SDK to take default API version from Oneview appliance
+- Refactored SDK to take default API version from Oneview appliance,
+- Added support for automatic publish of Docker Image when there is a new release in GitHub.
+
 ### Bug fixes & Enhancements:
 - [#29] (https://github.com/HewlettPackard/terraform-provider-oneview/issues/29) Add Description and Bios settings to server template