+++ CHANGELOG.md
@@ -5,9 +5,14 @@ This project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html
 - This release supports API2000 minimally where we can use OneView API2000 with this SDK. No new fields are added/deleted to support API2000.
 
 ### Oneview Features supported
+- Ethernet Network
+- FC Network
+- FCOE Network
 - Hypervisor Cluster Profile
 - Hypervisor Manager
+- Network Set
 - Server Certificates
+- Uplink Set
 
 ### Major changes:
 - Refactored SDK to take default API version from Oneview appliance,