+++ CHANGELOG.md
@@ -5,14 +5,27 @@ This project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html
 - This release supports API20000 minimally where we can use OneView API2000 with this SDK. No new fields are added/deleted to support API2000.
 
 ### Oneview Features supported
+- This release supports API2000 minimally where we can use OneView API2000 with this SDK. No new fields are added/deleted to support API2000.
+
+### Oneview Features supported
+- Ethernet Network
+- FC Network
+- FCOE Network
+- Hypervisor Cluster Profile
+- Hypervisor Manager
+- Network Set
+- Server Certificates
 - Storage Pool
 - Storage System
 - Storage Volume
 - Storage Volume Attachment
 - Storage Volume Template
+- Uplink Set
 
 ### Major changes:
-- Refactored SDK to take default API version from Oneview appliance
+- Refactored SDK to take default API version from Oneview appliance,
+- Added support for automatic publish of Docker Image when there is a new release in GitHub.
+
 ### Bug fixes & Enhancements:
 - [#29] (https://github.com/HewlettPackard/terraform-provider-oneview/issues/29) Add Description and Bios settings to server template