+++ CHANGELOG.md
@@ -2,6 +2,9 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html)
 # [v1.5.0] (unreleased)
 ### Notes
+- This release supports API20000 minimally where we can use OneView API2000 with this SDK. No new fields are added/deleted to support API2000.
+
+### Oneview Features supported
 - This release supports API2000 minimally where we can use OneView API2000 with this SDK. No new fields are added/deleted to support API2000.
 
 ### Oneview Features supported
@@ -12,6 +15,11 @@ This project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html
 - Hypervisor Manager
 - Network Set
 - Server Certificates
+- Storage Pool
+- Storage System
+- Storage Volume
+- Storage Volume Attachment
+- Storage Volume Template
 - Uplink Set
 
 ### Major changes: