+++ CHANGELOG.md
@@ -2,12 +2,27 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html)
 # [v1.5.0] (unreleased)
 ### Notes
-- This release supports API2000 minimally where we can use OneView API2000 with this SDK.
 
 ### Oneview Features supported
+- This release supports API2000 minimally where we can use OneView API2000 with this SDK. No new fields are added/deleted to support API2000.
+
+### Oneview Features supported
+- Ethernet Network
+- FC Network
+- FCOE Network
+- Hypervisor Cluster Profile
+- Hypervisor Manager
 - Logical Interconnect
+- Network Set
+- Server Certificates
 - Server Profile
 - Server Profile Template
+- Storage Pool
+- Storage System
+- Storage Volume
+- Storage Volume Attachment
+- Storage Volume Template
+- Uplink Set
 
 ### Major changes:
 - Refactored SDK to take default API version from Oneview appliance