+++ CHANGELOG.md
@@ -2,7 +2,6 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html)
 # [v1.5.0] (unreleased)
 ### Notes
-- This release supports API20000 minimally where we can use OneView API2000 with this SDK. No new fields are added/deleted to support API2000.
 
 ### Oneview Features supported
 - This release supports API2000 minimally where we can use OneView API2000 with this SDK. No new fields are added/deleted to support API2000.
@@ -13,8 +12,11 @@ This project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html
 - FCOE Network
 - Hypervisor Cluster Profile
 - Hypervisor Manager
+- Logical Interconnect
 - Network Set
 - Server Certificates
+- Server Profile
+- Server Profile Template
 - Storage Pool
 - Storage System
 - Storage Volume
@@ -23,7 +25,7 @@ This project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html
 - Uplink Set
 
 ### Major changes:
-- Refactored SDK to take default API version from Oneview appliance,
+- Refactored SDK to take default API version from Oneview appliance
 - Added support for automatic publish of Docker Image when there is a new release in GitHub.
 
 ### Bug fixes & Enhancements: