+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html)
-# [v1.5.0] (unreleased)
+# [v1.5.0]
 ### Notes
 - This release supports API2000 minimally where we can use OneView API2000 with this SDK. No new fields are added/deleted to support API2000.